restructuring of unstructured.py
