+++ resqpy/unstructured/__init__.py
@@ -1,0 +1,8 @@
+__all__ = ['UnstructuredGrid', 'HexaGrid', 'PrismGrid', 'VerticalPrismGrid', 'PyramidGrid',
+           'TetraGrid', 'valid_cell_shapes']
+
+from .unstructured_grid import UnstructuredGrid, valid_cell_shapes
+from .hexa_grid import HexaGrid
+from .prism_grid import PrismGrid, VerticalPrismGrid
+from .pyramid_grid import PyramidGrid
+from .tetra_grid import TetraGrid
